FLOAT

+++ core.py
@@ -289,7 +289,7 @@ def plot_transform(ct, size, esize, hits_course=9, hits_fine=65, custom=()):
 
     def draw(X, Y, color):
         xx, yy, _, _ = func(X, Y, zer, zer)
-        arr = np.array([xx, yy])
+        arr = np.array([xx, yy],dtype=np.float)
         print(arr)
         arr = arr[~np.isnan(arr).any(axis=1)]
         plt.plot(arr[0], arr[1], color)
